requirements.txt
requests
python_dotenv
prefect
mysql-connector-python